fix(merge): complete the upstream sync so the branch builds again
Both upstream-sync merges (e708a74d, 1ccf48c0) resolved conflicts by taking
one whole side per file. The result mixed upstream call sites with
branch-side definitions and no longer compiles:

  model/audit_log.go:223:26        user.AccessTokenCreatedAt undefined
  model/email_binding.go:111:13    undefined: lockNormalizedEmail
  model/option.go:60:78            undefined: constant.TaskPluginOverrideEnabled
  model/option.go:61:27            jsplugin.DefaultRegistry.SetOverrideEnabled undefined
  model/user_quota_adjustment.go   int vs int64 quota mismatch (x3)

Symbol provenance behind those errors:

  TaskPluginOverrideEnabled / SetOverrideEnabled
      5ca7ae79 x6/x8  ->  e708a74d x1/x1  ->  upstream x0
  AccessTokenCreatedAt / lockNormalizedEmail
      upstream x2/x2  ->  e708a74d x1/x1 (call site only)

The merges dropped the branch's own task-plugin override layer while keeping
its only call site, and kept upstream's audit_log.go/email_binding.go while
dropping the model/user.go definitions they call.

This commit completes the sync: the override layer is back with its
rebuild-on-candidate semantics, the access-token and normalized-email
definitions upstream depends on are restored, the wallet cap is carried as
int64 end to end, outbound HTTP and relay request validation are reconciled
with upstream, and the maintenance page regains the <SystemUpdateAction />
entry the merge removed.

All 63 files differ from 5ca7ae79 as well, so these are adaptations to
upstream main rather than reverts to the rc.34 blobs.

Gate: full Go suite, frontend suite, and the SQLite/MySQL/PostgreSQL
representative upgrade gate pass on exactly this tree.

+++ controller/responses_websocket_test.go
@@ -426,7 +426,7 @@ func assertResponsesWSAccounting(t *testing.T, fixture *responsesWSBillingTest, 
 	require.NoError(t, model.DB.First(fixture.user, fixture.user.Id).Error)
 	assert.Equal(t, 3000-charged, fixture.token.RemainQuota)
 	assert.Equal(t, charged, fixture.token.UsedQuota)
-	assert.Equal(t, 100000-charged, fixture.user.Quota)
+	assert.Equal(t, int64(100000-charged), fixture.user.Quota)
 	assert.Equal(t, charged, fixture.user.UsedQuota)
 }
 
@@ -521,7 +521,7 @@ func TestResponsesWebSocketReusesConnectionAndSettlesEachRequest(t *testing.T) {
 	require.NoError(t, model.DB.First(user, user.Id).Error)
 	assert.Zero(t, token.RemainQuota)
 	assert.Equal(t, 3000, token.UsedQuota)
-	assert.Equal(t, 97000, user.Quota)
+	assert.Equal(t, int64(97000), user.Quota)
 	assert.Equal(t, 3000, user.UsedQuota)
 }
