+++ service/quota_saturation_test.go
@@ -44,7 +44,7 @@ func TestAttachQuotaSaturationNestsUnderAdminInfo(t *testing.T) {
 	require.True(t, ok, "quota_saturation should be nested under admin_info")
 	require.Equal(t, "QuotaFromDecimal", sat["op"])
 	require.Equal(t, common.QuotaClampOverflow, sat["kind"])
-	require.Equal(t, common.MaxQuota, sat["clamped"])
+	require.Equal(t, int64(common.MaxQuota), sat["clamped"])
 }
 
 func TestCalcViolationFeeQuotaSaturates(t *testing.T) {
